Migrate `mark` module to submodule of `assignment`
Migrate `*Error` structs to `errors` module
