+++ tracker/core/src/class.rs
@@ -1,7 +1,7 @@
+use crate::errors::InvalidClassError::TotalValueOutOfRange;
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Debug, Display};
-use thiserror::Error;
 
 /// Generic representation a university/school class/paper/course.
 pub trait Classlike: Display + Debug + PartialEq + PartialOrd {
@@ -85,10 +85,7 @@ impl Classlike for Code {
 
     fn set_total_value(&mut self, value: f64) -> Result<()> {
         if !(0.0..=100.0).contains(&value) {
-            bail!(InvalidClassError::TotalValueOutOfRange(
-                self.name().to_owned(),
-                value
-            ));
+            bail!(TotalValueOutOfRange(self.name().to_owned(), value));
         }
 
         trace!("{self} -> Total value -> {value}");
@@ -144,10 +141,7 @@ impl Classlike for Class {
 
     fn set_total_value(&mut self, value: f64) -> Result<()> {
         if !(0.0..=100.0).contains(&value) {
-            bail!(InvalidClassError::TotalValueOutOfRange(
-                self.name().to_owned(),
-                value
-            ));
+            bail!(TotalValueOutOfRange(self.name().to_owned(), value));
         }
 
         trace!("{self} -> Total value -> {value}");
@@ -166,12 +160,4 @@ impl Default for Class {
     fn default() -> Self {
         Self::with_name("DEFAULT", "Default Class")
     }
-}
-
-/// The [class](Classlike) is invalid.
-#[derive(Error, Debug)]
-pub enum InvalidClassError {
-    /// The total value of all assignments must be within `0.0..=100.0`
-    #[error("{0} -> Total value ({1}) must be within 0.0..=100.0")]
-    TotalValueOutOfRange(String, f64),
 }
